Clean up namespaces
