+++ src/functionalTest/java/uk/gov/hmcts/reform/draftstore/data/PaginationTest.java
@@ -8,7 +8,7 @@ import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import uk.gov.hmcts.reform.draftstore.data.model.Draft;
-import uk.gov.hmcts.reform.draftstore.endpoint.v3.helpers.SampleData;
+import uk.gov.hmcts.reform.draftstore.controllers.helpers.SampleData;
 
 import java.util.List;
 import java.util.function.Supplier;
