+++ src/main/java/uk/gov/hmcts/reform/draftstore/controllers/EndpointExceptionHandler.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.draftstore.endpoint;
+package uk.gov.hmcts.reform.draftstore.controllers;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -12,7 +12,7 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.client.HttpClientErrorException;
 import org.springframework.web.context.request.WebRequest;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
-import uk.gov.hmcts.reform.draftstore.endpoint.domain.ErrorResult;
+import uk.gov.hmcts.reform.draftstore.domain.ErrorResult;
 import uk.gov.hmcts.reform.draftstore.exception.AuthorizationException;
 import uk.gov.hmcts.reform.draftstore.exception.NoDraftFoundException;
 import uk.gov.hmcts.reform.draftstore.service.crypto.InvalidKeyException;
@@ -29,11 +29,11 @@ import static org.springframework.http.HttpStatus.BAD_REQUEST;
 import static org.springframework.http.HttpStatus.FORBIDDEN;
 import static org.springframework.http.HttpStatus.INTERNAL_SERVER_ERROR;
 import static org.springframework.http.HttpStatus.NOT_FOUND;
-import static uk.gov.hmcts.reform.draftstore.endpoint.domain.ErrorCode.BAD_ARGUMENT;
-import static uk.gov.hmcts.reform.draftstore.endpoint.domain.ErrorCode.INVALID_AUTH_TOKEN;
-import static uk.gov.hmcts.reform.draftstore.endpoint.domain.ErrorCode.NO_RECORD_FOUND;
-import static uk.gov.hmcts.reform.draftstore.endpoint.domain.ErrorCode.SERVER_ERROR;
-import static uk.gov.hmcts.reform.draftstore.endpoint.domain.ErrorCode.USER_DETAILS_SERVICE_ERROR;
+import static uk.gov.hmcts.reform.draftstore.domain.ErrorCode.BAD_ARGUMENT;
+import static uk.gov.hmcts.reform.draftstore.domain.ErrorCode.INVALID_AUTH_TOKEN;
+import static uk.gov.hmcts.reform.draftstore.domain.ErrorCode.NO_RECORD_FOUND;
+import static uk.gov.hmcts.reform.draftstore.domain.ErrorCode.SERVER_ERROR;
+import static uk.gov.hmcts.reform.draftstore.domain.ErrorCode.USER_DETAILS_SERVICE_ERROR;
 
 @ControllerAdvice
 public class EndpointExceptionHandler extends ResponseEntityExceptionHandler {
